- Ajout d'une option de sauvegarde SAVE_SIMULATION

+++ main.py
@@ -34,7 +34,8 @@ def show_var_infos(vari,title="swow_var_info", dirpath = None):
     fig,ax = plt.subplots()
     ax.plot(vari)
     ax.set_title("variance evolution from t=0 to t=T")
-    fig.savefig(dirpath+"/"+ title_file +".png")
+    if dirpath != None:
+        fig.savefig(dirpath+"/"+ title_file +".png")
     plt.show()
 
 env = gym.make('mopsi-env-v0')
@@ -42,15 +43,17 @@ env = gym.make('mopsi-env-v0')
 # Configuration
 
 env.config["number_of_lane"] = 1
-env.config["other_vehicles"] = 19
+env.config["other_vehicles"] = 0
 env.config["controlled_vehicles"] = 1
-env.config["duration"] = 90
+env.config["duration"] = 1000
 
 
 env.config["screen_width"] = 1000
 env.config["screen_height"] = 1000
 
-env.reset()
+SAVE_SIMULATION = False
+
+env.reset("rl")
 
 # Main program
 
@@ -67,14 +70,15 @@ if __name__ == "__main__":
         raise SystemError("Simulation must have at least 100it")
 
     # Initialisation du dossier résultats
-    time = str(datetime.now().date()) + "___" + str(
+    if SAVE_SIMULATION:
+        time = str(datetime.now().date()) + "___" + str(
         datetime.now().hour) + "_" + str(datetime.now().minute) + "_" + str(datetime.now().second)
-    result_folder_path = "results/Simulation__"+ time
-    os.mkdir(result_folder_path)
+        result_folder_path = "results/Simulation__"+ time
+        os.mkdir(result_folder_path)
 
-    time_gif = duration//2
-    duration_gif = 50
-    filenames = []
+        time_gif = duration//2
+        duration_gif = 50
+        filenames = []
 
 
     # Boucle principale
@@ -82,9 +86,10 @@ if __name__ == "__main__":
 
         # Action
         obs, reward, done, info = env.step([0,0])
+        env.render()
 
-        if i>= time_gif and i<=time_gif + duration_gif :
-        # Rendu graphique ( à commenter pour accelerer le code )
+        # Gif
+        if SAVE_SIMULATION and i>= time_gif and i<=time_gif + duration_gif:
             name_picture = result_folder_path + "/" + "render"+str(i)+".png"
             filenames.append(name_picture)
             plt.imsave(name_picture,env.render(mode="rgb_array"))
@@ -95,14 +100,17 @@ if __name__ == "__main__":
 
     # create gif
 
-    with imageio.get_writer(result_folder_path + "/results.gif", mode = "I") as writer :
-        for filename in filenames:
-            image = imageio.imread(filename)
-            writer.append_data(image)
-    for filename in set(filenames):
-        os.remove(filename)
+    if SAVE_SIMULATION:
+        with imageio.get_writer(result_folder_path + "/results.gif", mode = "I") as writer :
+            for filename in filenames:
+                image = imageio.imread(filename)
+                writer.append_data(image)
+        for filename in set(filenames):
+            os.remove(filename)
 
-    show_var_infos(hist[10:], "traffic_" + str(real_nb_vehicles) + "_vehicles" + "__" + str(duration) + "it__",
+        show_var_infos(hist[10:], "traffic_" + str(real_nb_vehicles) + "_vehicles" + "__" + str(duration) + "it__",
                    dirpath=result_folder_path)
 
-
+    else :
+        show_var_infos(hist[10:], "traffic_" + str(real_nb_vehicles) + "_vehicles" + "__" + str(duration) + "it__",
+                   dirpath=None)
